removing old nav
